+++ contracts/ETHMixer.sol
@@ -17,7 +17,7 @@ contract ETHMixer is Mixer {
   constructor(
     IVerifier _verifier,
     uint256 _denomination,
-    uint8 _merkleTreeHeight,
+    uint32 _merkleTreeHeight,
     address _operator
   ) Mixer(_verifier, _denomination, _merkleTreeHeight, _operator) public {
   }
